docs(refactor): use default NavBar & swizzle our own color switch
